use new chainreader methods
