+++ packages/txservice/src/chainreader.ts
@@ -1,10 +1,20 @@
-import { Signer, providers, BigNumber } from "ethers";
+import { Signer, providers, BigNumber, constants } from "ethers";
 import { createLoggingContext, Logger, RequestContext } from "@connext/nxtp-utils";
 
 import { TransactionServiceConfig, validateTransactionServiceConfig, DEFAULT_CONFIG, ChainConfig } from "./config";
 import { ReadTransaction } from "./types";
 import { ChainRpcProvider } from "./provider";
-import { ConfigurationError, ProviderNotConfigured } from "./error";
+import { ChainNotSupported, ConfigurationError, ProviderNotConfigured } from "./error";
+import { getDeployedChainIdsForGasFee, getDeployedPriceOracleContract, getPriceOracleInterface } from "./contracts";
+
+export const GAS_ESTIMATES = {
+  prepare: "112000", // https://etherscan.io/tx/0x4e1be107ca80265b7ae65f77e00f14dd726d08dae763955fb1cf2a754d9cc1a8 example tx, add 5% buffer
+  fulfill: "126000", // https://bscscan.com/tx/0xcaba240ab17f006586086cd460fffab09a028905d7c497c31667c1d5eb58e153 example tx, add 5% buffer
+  prepareL1: "17300", // https://optimistic.etherscan.io/tx/0xd3ae8d8980aa464c4256ef6c734f7eb58211a02a6016201903e30fd35ec3bff8
+  fulfillL1: "11800", // https://optimistic.etherscan.io/tx/0x280a1e70c10095d748babb85fa56fdf8285cdcae3e3962eae3dc451045c0b220
+};
+
+const NO_ORACLE_CHAINS: number[] = [];
 
 // TODO: I do not like that this is generally a passthrough class now - all it handles is the mapping. We should
 // probably just expose a provider getter method and have the consumer call that to access the target ChainRpcProvider
@@ -37,7 +47,7 @@ export class ChainReader {
     this.setupProviders(requestContext, signer);
   }
 
-  /// INTERFACE
+  /// CHAIN READING METHODS
   /**
    * Create a non-state changing contract call. Returns hexdata that needs to be decoded.
    *
@@ -132,6 +142,216 @@ export class ChainReader {
     return await this.getProvider(chainId).getCode(address);
   }
 
+  /// CONTRACT READ METHODS
+  /**
+   * Gets token price in usd from price oracle
+   *
+   * @param chainId - The network identifier.
+   * @param assetId - The asset address to get price for.
+   */
+  async getTokenPrice(chainId: number, assetId: string, _requestContext?: RequestContext): Promise<BigNumber> {
+    const { requestContext } = createLoggingContext(this.getTokenPrice.name, _requestContext);
+    const priceOracleContract = this.getPriceOracleContract(chainId, requestContext);
+    const encodedTokenPriceData = getPriceOracleInterface().encodeFunctionData("getTokenPrice", [assetId]);
+    const tokenPrice = await this.readTx({
+      chainId,
+      to: priceOracleContract.address,
+      data: encodedTokenPriceData,
+    });
+    return BigNumber.from(tokenPrice);
+  }
+
+  /**
+   * Helper to calculate router gas fee in token for prepare
+   *
+   * @param sendingAssetId The asset address on source chain
+   * @param sendingChainId The source chain Id
+   * @param receivingAssetId The asset address on destination chain
+   * @param receivingChainId The destination chain Id
+   * @param _outputDecimals Decimal number of receiving asset
+   * @param requestContext Request context instance
+   */
+  async calculateGasFeeInReceivingTokenForPrepare(
+    receivingChainId: number,
+    receivingAssetId: string,
+    outputDecimals: number,
+    sendingChainId?: number,
+    sendingAssetId?: string,
+    _requestContext?: RequestContext,
+  ): Promise<BigNumber> {
+    const { requestContext, methodContext } = createLoggingContext(
+      this.calculateGasFeeInReceivingTokenForPrepare.name,
+      _requestContext,
+    );
+    this.logger.info("Method start", requestContext, methodContext, {
+      sendingChainId,
+      sendingAssetId,
+      receivingAssetId,
+      receivingChainId,
+      outputDecimals,
+    });
+
+    const chaindIdsForGasFee = getDeployedChainIdsForGasFee();
+
+    if (
+      sendingChainId !== undefined &&
+      !chaindIdsForGasFee.includes(sendingChainId) &&
+      !chaindIdsForGasFee.includes(receivingChainId)
+    ) {
+      return constants.Zero;
+    }
+    let totalCost = constants.Zero;
+    // TODO: this is returning zero when doing a rinkeby to goerli tx. i believe this is because the oracle
+    // is not configured for goerli so theres no way to translate the price to goerli
+    // TODO: we can combine these into just 2 if statements and remove the repeated logic
+    // calculate receiving token amount for gas fee
+    // if chaindIdsForGasFee includes only sendingChainId, calculate gas fee for fulfill transactions
+    // if chaindIdsForGasFee includes only receivingChainId, calculate gas fee for prepare transactions
+
+    const tokenPricingSendingChain = sendingChainId && NO_ORACLE_CHAINS.includes(sendingChainId) ? 1 : sendingChainId;
+    const tokenPricingReceivingChain = NO_ORACLE_CHAINS.includes(receivingChainId) ? 1 : receivingChainId;
+
+    this.logger.info("Getting token prices", requestContext, methodContext, {
+      tokenPricingSendingChain,
+      tokenPricingReceivingChain,
+      sendingAssetId,
+      receivingAssetId,
+      outputDecimals,
+    });
+    if (chaindIdsForGasFee.includes(sendingChainId)) {
+      const gasLimitForFulfill = BigNumber.from(GAS_ESTIMATES.fulfill);
+      const [ethPriceInSendingChain, receivingTokenPrice, gasPriceInSendingChain] = await Promise.all([
+        this.getTokenPrice(tokenPricingSendingChain, constants.AddressZero),
+        this.getTokenPrice(tokenPricingSendingChain, sendingAssetId),
+        this.getGasPrice(sendingChainId, requestContext),
+      ]);
+
+      // https://community.optimism.io/docs/users/fees-2.0.html#fees-in-a-nutshell
+      let l1GasInUsd = BigNumber.from(0);
+      if (sendingChainId === 10) {
+        const gasPriceMainnet = await this.getGasPrice(1, requestContext);
+        l1GasInUsd = gasPriceMainnet.mul(GAS_ESTIMATES.fulfillL1).mul(ethPriceInSendingChain);
+      }
+      const gasAmountInUsd = gasPriceInSendingChain.mul(gasLimitForFulfill).mul(ethPriceInSendingChain).add(l1GasInUsd);
+      const tokenAmountForGasFee = receivingTokenPrice.isZero()
+        ? constants.Zero
+        : gasAmountInUsd.div(receivingTokenPrice).div(BigNumber.from(10).pow(18 - outputDecimals));
+
+      totalCost = totalCost.add(tokenAmountForGasFee);
+      this.logger.info("Calculated cost on sending chain", requestContext, methodContext, {
+        totalCost: totalCost.toString(),
+        l1GasInUsd: l1GasInUsd.toString(),
+        ethPriceInSendingChain: ethPriceInSendingChain.toString(),
+        receivingTokenPrice: receivingTokenPrice.toString(),
+        gasPriceInSendingChain: gasPriceInSendingChain.toString(),
+      });
+    }
+
+    if (chaindIdsForGasFee.includes(receivingChainId)) {
+      const gasLimitForPrepare = BigNumber.from(GAS_ESTIMATES.prepare);
+      const [ethPriceInReceivingChain, receivingTokenPrice, gasPriceInReceivingChain] = await Promise.all([
+        this.getTokenPrice(tokenPricingReceivingChain, constants.AddressZero),
+        this.getTokenPrice(tokenPricingReceivingChain, receivingAssetId),
+        this.getGasPrice(receivingChainId, requestContext),
+      ]);
+
+      // https://community.optimism.io/docs/users/fees-2.0.html#fees-in-a-nutshell
+      let l1GasInUsd = BigNumber.from(0);
+      if (receivingChainId === 10) {
+        const gasPriceMainnet = await this.getGasPrice(1, requestContext);
+        l1GasInUsd = gasPriceMainnet.mul(GAS_ESTIMATES.prepareL1).mul(ethPriceInReceivingChain);
+      }
+      const gasAmountInUsd = gasPriceInReceivingChain
+        .mul(gasLimitForPrepare)
+        .mul(ethPriceInReceivingChain)
+        .add(l1GasInUsd);
+      const tokenAmountForGasFee = receivingTokenPrice.isZero()
+        ? constants.Zero
+        : gasAmountInUsd.div(receivingTokenPrice).div(BigNumber.from(10).pow(18 - outputDecimals));
+
+      totalCost = totalCost.add(tokenAmountForGasFee);
+      this.logger.info("Calculated cost on receiving chain", requestContext, methodContext, {
+        totalCost: totalCost.toString(),
+        tokenAmountForGasFee: tokenAmountForGasFee.toString(),
+        l1GasInUsd: l1GasInUsd.toString(),
+        ethPriceInSendingChain: ethPriceInReceivingChain.toString(),
+        receivingTokenPrice: receivingTokenPrice.toString(),
+        gasPriceInSendingChain: gasPriceInReceivingChain.toString(),
+      });
+    }
+
+    // convert back to the intended decimals
+    return totalCost;
+  }
+
+  /**
+   * Helper to calculate router gas fee in token for meta transaction
+   *
+   * @param receivingAssetId The asset address on destination chain
+   * @param receivingChainId The destination chain Id
+   * @param outputDecimals Decimal number of receiving asset
+   * @param requestContext Request context instance
+   */
+  async calculateGasFeeInReceivingTokenForFulfill(
+    receivingChainId: number,
+    receivingAssetId: string,
+    outputDecimals: number,
+    _requestContext: RequestContext,
+  ): Promise<BigNumber> {
+    const { requestContext, methodContext } = createLoggingContext(
+      this.calculateGasFeeInReceivingTokenForFulfill.name,
+      _requestContext,
+    );
+    this.logger.info("Method start", requestContext, methodContext, {
+      receivingChainId,
+      receivingAssetId,
+      outputDecimals,
+    });
+    const chaindIdsForGasFee = getDeployedChainIdsForGasFee();
+
+    if (!chaindIdsForGasFee.includes(receivingChainId)) return constants.Zero;
+    let totalCost = constants.Zero;
+
+    const tokenPricingReceivingChain = NO_ORACLE_CHAINS.includes(receivingChainId) ? 1 : receivingChainId;
+
+    if (chaindIdsForGasFee.includes(receivingChainId)) {
+      const gasLimitForFulfill = BigNumber.from(GAS_ESTIMATES.fulfill);
+      const [ethPriceInReceivingChain, receivingTokenPrice, gasPriceInReceivingChain] = await Promise.all([
+        this.getTokenPrice(tokenPricingReceivingChain, constants.AddressZero),
+        this.getTokenPrice(tokenPricingReceivingChain, receivingAssetId),
+        this.getGasPrice(receivingChainId, requestContext),
+      ]);
+
+      // https://community.optimism.io/docs/users/fees-2.0.html#fees-in-a-nutshell
+      let l1GasInUsd = BigNumber.from(0);
+      if (receivingChainId === 10) {
+        const gasPriceMainnet = await this.getGasPrice(1, requestContext);
+        l1GasInUsd = gasPriceMainnet.mul(GAS_ESTIMATES.prepareL1).mul(ethPriceInReceivingChain);
+      }
+
+      const gasAmountInUsd = gasPriceInReceivingChain
+        .mul(gasLimitForFulfill)
+        .mul(ethPriceInReceivingChain)
+        .add(l1GasInUsd);
+      const tokenAmountForGasFee = receivingTokenPrice.isZero()
+        ? constants.Zero
+        : gasAmountInUsd.div(receivingTokenPrice).div(BigNumber.from(10).pow(18 - outputDecimals));
+
+      this.logger.info("Calculated cost on receiving chain for fulfill", requestContext, methodContext, {
+        totalCost: totalCost.toString(),
+        tokenAmountForGasFee: tokenAmountForGasFee.toString(),
+        l1GasInUsd: l1GasInUsd.toString(),
+        ethPriceInReceivingChain: ethPriceInReceivingChain.toString(),
+        receivingTokenPrice: receivingTokenPrice.toString(),
+        gasPriceInReceivingChain: gasPriceInReceivingChain.toString(),
+      });
+
+      totalCost = totalCost.add(tokenAmountForGasFee);
+    }
+
+    return totalCost;
+  }
+
   /// HELPERS
   /**
    * Helper to wrap getting provider for specified chain ID.
@@ -175,5 +395,13 @@ export class ChainReader {
       const provider = new ChainRpcProvider(this.logger, chainIdNumber, chain, this.config, signer);
       this.providers.set(chainIdNumber, provider);
     });
+  }
+
+  private getPriceOracleContract(chainId: number, requestContext: RequestContext): { address: string; abi: any } {
+    const priceOracleContract = getDeployedPriceOracleContract(chainId);
+    if (!priceOracleContract) {
+      throw new ChainNotSupported(chainId.toString(), requestContext);
+    }
+    return priceOracleContract;
   }
 }
